+++ example/storybook-nativewind/src/core-components/nativewind/actionsheet/index.tsx
@@ -78,6 +78,7 @@ const ItemWrapper = React.forwardRef(({ ...props }: any, ref?: any) => {
 });
 
 const AnimatedPressable = createMotionAnimatedComponent(Pressable);
+
 export const UIActionsheet = createActionsheet({
   Root: View,
   Content: Motion.View,