+++ example/storybook-nativewind/src/core-components/nativewind/heading/styles.tsx
@@ -1,5 +1,4 @@
 import { tva } from '@gluestack-ui/nativewind-utils/tva';
-
 import { isWeb } from '@gluestack-ui/nativewind-utils/IsWeb';
 const baseStyle = isWeb
   ? 'font-sans tracking-sm bg-transparent border-0 box-border display-inline list-none margin-0 padding-0 position-relative text-start no-underline whitespace-pre-wrap word-wrap-break-word'