+++ packages/unstyled/alert/src/types.ts
@@ -1,18 +1,22 @@
+import React from 'react';
 import type { ViewProps } from 'react-native';
 
 export interface InterfaceAlertProps extends ViewProps {
-  children?: JSX.Element | Array<JSX.Element>;
+  children?: React.ReactNode;
 }
 
 export type IAlertComponentType<StyledAlert, StyledAlertText, StyledAlertIcon> =
   React.ForwardRefExoticComponent<
-    React.RefAttributes<StyledAlert> & StyledAlert & InterfaceAlertProps
+    React.RefAttributes<StyledAlert> &
+      React.PropsWithoutRef<StyledAlert & InterfaceAlertProps>
   > & {
     Text: React.ForwardRefExoticComponent<
-      React.RefAttributes<StyledAlertText> & StyledAlertText
+      React.RefAttributes<StyledAlertText> &
+        React.PropsWithoutRef<StyledAlertText>
     >;
     Icon: React.ForwardRefExoticComponent<
-      React.RefAttributes<StyledAlertIcon> & StyledAlertIcon
+      React.RefAttributes<StyledAlertIcon> &
+        React.PropsWithoutRef<StyledAlertIcon>
     >;
   };