+++ packages/unstyled/input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gluestack-ui/input
 
+## 0.1.31
+
+### Patch Changes
+
+- fix: ref typing for all the components
+- Updated dependencies
+  - @gluestack-ui/form-control@0.1.18
+
 ## 0.1.30
 
 ### Patch Changes